refactor: added domain for users, modified properties and gradle, created missing entity UserRole
