+++ backend/src/main/java/com/polishbank/bank_a/BankAApplication.java
@@ -2,8 +2,10 @@ package com.polishbank.bank_a;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.scheduling.annotation.EnableScheduling;
 
 @SpringBootApplication
+@EnableScheduling
 public class BankAApplication {
 
 	public static void main(String[] args) {
