+++ backend/src/main/java/com/polishbank/bank_a/domain/user/UserRepository.java
@@ -1,0 +1,12 @@
+package com.polishbank.bank_a.domain.user;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+import java.util.UUID;
+
+public interface UserRepository extends JpaRepository<User, UUID> {
+    Optional<User> findByEmail(String email);
+    boolean existsByEmail(String email);
+    boolean existsByCustomerNumber(String customerNumber);
+}
