+++ backend/src/main/java/com/polishbank/bank_a/domain/user/UserRole.java
@@ -1,0 +1,5 @@
+package com.polishbank.bank_a.domain.user;
+
+public enum UserRole {
+    CUSTOMER, ADMIN, PARENT, JUNIOR
+}
