--- backend/src/main/java/com/polishbank/bank_a/entity/User.java
@@ -1,30 +1,0 @@
-package com.polishbank.bank_a.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-import java.time.LocalDate;
-import java.util.UUID;
-
-@Entity
-@Table(name = "users")
-@Data
-@NoArgsConstructor
-@AllArgsConstructor
-@Builder
-public class User {
-    @Id
-    @GeneratedValue(strategy = GenerationType.UUID)
-    private UUID id;
-    
-    private String customerNumber;
-    
-    private String firstName;
-    private String lastName;
-    private String email;
-    private String passwordHash;
-    private String phoneNumber;
-    
-    private LocalDate dateOfBirth;
-    
-    private String role;
-}
