+++ backend/src/main/resources/application.properties
@@ -3,3 +3,21 @@ spring.flyway.enabled=true
 spring.flyway.baseline-on-migrate=true
 spring.flyway.check-location=false
 logging.level.org.flywaydb=DEBUG
+
+spring.datasource.url=${SPRING_DATASOURCE_URL:jdbc:postgresql://localhost:5432/bank_a}
+spring.datasource.username=${SPRING_DATASOURCE_USERNAME:bank_admin}
+spring.datasource.password=${SPRING_DATASOURCE_PASSWORD:secretpassword}
+spring.jpa.hibernate.ddl-auto=validate
+spring.jpa.show-sql=false
+
+app.jwt.secret=${JWT_SECRET:change-me-64-char-min-secret-key-placeholder-here!!}
+app.jwt.expiration-ms=86400000
+
+integration.elixir.base-url=${ELIXIR_BASE_URL:http://elixir-service:8080}
+app.elixir.poll-interval-ms=30000
+integration.klik.base-url=${KLIK_BASE_URL:https://api.klik.example.com/api/v1}
+integration.klik.api-key=${KLIK_API_KEY:dev-key}
+integration.klik.zone=PL
+integration.cards.grpc-host=${CARDS_GRPC_HOST:card-provider-service}
+integration.cards.grpc-port=${CARDS_GRPC_PORT:50051}
+integration.cards.gateway-url=${CARDS_GATEWAY_URL:http://payment-gateway-service:8000}
